Complete rewrite of the code
Lots of improvements, when I wrote this bot I didn't knew lots of
things, so the code was far from pythonic.

This rewrite also standardize the way modules should be, and in a future
commit I intend to create a exemple module file and load the modules
automatically at startup, maybe even using inotify to detect file
changes and reload on the go.